refac) style 컴포넌트 관심사 분리
