+++ src/App.jsx
@@ -8,8 +8,6 @@ import {
 
 import { useDispatch } from 'react-redux';
 
-import styled from '@emotion/styled';
-
 import HomePage from './HomePage';
 import AboutPage from './AboutPage';
 import LoginPage from './LoginPage';
@@ -17,30 +15,12 @@ import RestaurantsPage from './RestaurantsPage';
 import RestaurantPage from './RestaurantPage';
 import NotFoundPage from './NotFoundPage';
 
+import Header from './styles/Header';
+import Container from './styles/Container';
+
 import { setAccessToken } from './slice';
 
 import { loadItem } from './services/storage';
-
-const Container = styled.div({
-  margin: '0 auto',
-  width: '90%',
-});
-
-const Header = styled.header({
-  backgroundColor: '#EEE',
-  '& h1': {
-    fontSize: '1.5em',
-    margin: 0,
-    padding: '1em .5em',
-  },
-  '& a': {
-    color: '#555',
-    textDecoration: 'none',
-    '&:hover': {
-      color: '#000',
-    },
-  },
-});
 
 export default function App() {
   const dispatch = useDispatch();
