+++ src/HomePage.jsx
@@ -2,43 +2,21 @@ import React from 'react';
 
 import { Link } from 'react-router-dom';
 
-import styled from '@emotion/styled';
+import Title from './styles/Title';
+import NavList from './styles/NavList';
+import NavItem from './styles/NavItem';
 
-const Title = styled.h2({
-  fontSize: '2em',
-  margin: 0,
-  padding: '.4em 0',
-});
-
-const List = styled.ul({
-  display: 'flex',
-  margin: 0,
-  padding: 0,
-  listStyle: 'none',
-});
-
-const Item = styled.li({
-  marginRight: '1em',
-  '& a': {
-    color: '#333',
-    textDecoration: 'none',
-    '&:hover': {
-      fontWeight: 'bold',
-      color: '#000',
-    },
-  },
-});
 
 export default function HomePage() {
   return (
     <div>
       <Title>Home</Title>
-      <List>
-        <Item><Link to="/about">About</Link></Item>
-        <Item><Link to="/login">Log in</Link></Item>
-        <Item><Link to="/restaurants">Restaurants</Link></Item>
-        <Item><Link to="/xxx">멸망의 길</Link></Item>
-      </List>
+      <NavList>
+        <NavItem><Link to="/about">About</Link></NavItem>
+        <NavItem><Link to="/login">Log in</Link></NavItem>
+        <NavItem><Link to="/restaurants">Restaurants</Link></NavItem>
+        <NavItem><Link to="/xxx">멸망의 길</Link></NavItem>
+      </NavList>
     </div>
   );
 }
