+++ src/styles/Container.jsx
@@ -1,0 +1,10 @@
+import styled from '@emotion/styled';
+
+const Container = styled.div({
+  display: 'flex',
+  justifyContent: 'space-between',
+  height: '100vh',
+  width: '100%',
+});
+
+export default Container;
