+++ src/styles/NavItem.jsx
@@ -1,0 +1,15 @@
+import styled from '@emotion/styled';
+
+const NavItem = styled.li({
+  marginRight: '1em',
+  '& a': {
+    color: '#333',
+    textDecoration: 'none',
+    '&:hover': {
+      fontWeight: 'bold',
+      color: '#000',
+    },
+  },
+});
+
+export default NavItem;
