+++ src/styles/NavList.jsx
@@ -1,0 +1,10 @@
+import styled from '@emotion/styled';
+
+const NavList = styled.ul({
+  display: 'flex',
+  margin: 0,
+  padding: 0,
+  listStyle: 'none',
+});
+
+export default NavList;
