+++ src/styles/Title.jsx
@@ -1,0 +1,9 @@
+import styled from '@emotion/styled';
+
+const Title = styled.h2({
+  fontSize: '2em',
+  margin: 0,
+  padding: '.4em 0',
+});
+
+export default Title;
